+++ src/main/kotlin/taehyeon/brothers/matchreal/presentation/daily/dto/request/TagRequest.kt
@@ -1,0 +1,9 @@
+package taehyeon.brothers.matchreal.presentation.daily.dto.request
+
+data class TagAddRequest(
+    val tagName: String,
+)
+
+data class TagRemoveRequest(
+    val tagId: Long,
+)
