+++ src/main/kotlin/taehyeon/brothers/matchreal/presentation/daily/dto/response/TagResponse.kt
@@ -1,0 +1,6 @@
+package taehyeon.brothers.matchreal.presentation.daily.dto.response
+
+data class AddTagResponse(
+    val tagId: Long,
+    val tagName: String,
+)
